We have pictures on iOS. Not rendering with right view yet, need to investigate screen coords.  But, we have our wonderful yellow cube on the screen... sort of.
Had to fix some remaining issues with thread_local.h hacks.
And shaders needed mediump float declaration.

+++ src/core/platutils_ios.cpp
@@ -58,8 +58,8 @@ void init_time()
 
 f32 now()
 {
-    ASSERT_MSG(sStartTimeAbsolute != 0, "init_time_utils must be called first");
-    ASSERT_MSG(sTimebaseDenom != 0, "init_time_utils must be called first");
+    ASSERT_MSG(sStartTimeAbsolute != (u64)0, "init_time_utils must be called first");
+    ASSERT_MSG(sTimebaseDenom != (u32)0, "init_time_utils must be called first");
 
     u64 nowAbsolute = mach_absolute_time();
     u64 deltaAbsolute = nowAbsolute - sStartTimeAbsolute;
